sca-os/requirements.txt
jsonschema
pyyaml

